requirements.txt
blinker==1.7.0
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.0
identify==2.5.33
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
nodeenv==1.8.0
packaging==23.2
platformdirs==4.1.0
pluggy==1.3.0
pre-commit==3.6.0
psycopg2-binary==2.9.9
pytest==7.4.3
PyYAML==6.0.1
tomli==2.0.1
virtualenv==20.25.0
Werkzeug==3.0.1
